refactor: 위치 변경 및 record 적용
