--- ai-service/src/main/java/com/ojosama/moderation/domain/event/payload/AiReportEvent.java
@@ -1,12 +1,0 @@
-package com.ojosama.moderation.domain.event.payload;
-
-import java.util.UUID;
-
-public record AiReportEvent(
-    UUID targetId,
-    UUID targetUserId,
-    String targetType,
-    String category,
-    String description,
-    String content
-) { }
