+++ operation-service/src/main/java/com/ojosama/report/domain/event/payload/AiReportEvent.java
@@ -1,0 +1,12 @@
+package com.ojosama.report.domain.event.payload;
+
+import java.util.UUID;
+
+public record AiReportEvent(
+        UUID targetId,
+        UUID targetUserId,
+        String targetType,
+        String category,
+        String description,
+        String content
+) { }
